requirements.txt
httpx>=0.27.0
pydantic>=2.8.0
mcp>=1.2.0
anyio>=4.4.0
python-dotenv>=1.0.1
google-generativeai>=0.7.2


